+++ app/build.gradle
@@ -13,6 +13,7 @@ dependencies {
     implementation project(':feature-pokedex')
     implementation project(':core:network')
     implementation project(':core:data')
+    implementation project(':core:domain')
 
     implementation androidx_appcompat
     implementation material_design
