+++ core/domain/build.gradle
@@ -1,0 +1,8 @@
+apply from:"../../kotlin-module.gradle"
+
+dependencies {
+    implementation project(':core:model')
+    implementation project(':core:data')
+
+    implementation kodein
+}
