+++ feature-pokedex/build.gradle
@@ -10,6 +10,7 @@ dependencies {
     implementation project(':core:model')
     implementation project(':core:network')
     implementation project(':core:data')
+    implementation project(':core:domain')
     implementation project(':features-common')
 
     implementation kodein
