+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/di/pokedexModule.kt
@@ -1,7 +1,5 @@
 package com.victor.feature_pokedex.di
 
-import com.victor.feature_pokedex.domain.PokemonInformationUseCase
-import com.victor.feature_pokedex.domain.PokemonListUseCase
 import com.victor.feature_pokedex.presentation.ui.details.DetailsViewModel
 import com.victor.feature_pokedex.presentation.ui.home.HomeViewModel
 import org.kodein.di.DI
@@ -10,14 +8,6 @@ import org.kodein.di.instance
 import org.kodein.di.provider
 
 val pokedexModule = DI.Module("pokedex") {
-
-    bind<PokemonListUseCase>() with provider {
-        PokemonListUseCase(instance(), instance())
-    }
-
-    bind<PokemonInformationUseCase>() with provider {
-        PokemonInformationUseCase(instance(), instance())
-    }
 
     bind<HomeViewModel>() with provider {
         HomeViewModel(instance(), instance(), instance())
