+++ feature-pokedex/src/main/java/com/victor/feature_pokedex/presentation/ui/details/DetailsViewModel.kt
@@ -2,7 +2,7 @@ package com.victor.feature_pokedex.presentation.ui.details
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import com.victor.feature_pokedex.domain.PokemonInformationUseCase
+import com.victor.domain.PokemonInformationUseCase
 import com.victor.features_common.State
 import com.victor.features_common.manageStateDuringRequest
