+++ settings.gradle
@@ -12,3 +12,4 @@ include ':features-common'
 include ':core:model'
 include ':core:network'
 include ':core:data'
+include ':core:domain'
